begin slice
