+++ src/main/java/org/apache/sysds/runtime/compress/AbstractCompressedMatrixBlock.java
@@ -31,7 +31,6 @@ import org.apache.sysds.runtime.DMLRuntimeException;
 import org.apache.sysds.runtime.compress.colgroup.ColGroup;
 import org.apache.sysds.runtime.compress.colgroup.ColGroupUncompressed;
 import org.apache.sysds.runtime.compress.colgroup.ColGroupValue;
-import org.apache.sysds.runtime.controlprogram.caching.CacheBlock;
 import org.apache.sysds.runtime.controlprogram.caching.MatrixObject.UpdateType;
 import org.apache.sysds.runtime.instructions.cp.CM_COV_Object;
 import org.apache.sysds.runtime.instructions.cp.ScalarObject;
@@ -221,26 +220,6 @@ public abstract class AbstractCompressedMatrixBlock extends MatrixBlock {
 		printDecompressWarning("leftIndexingOperations");
 		MatrixBlock tmp = decompress();
 		return tmp.leftIndexingOperations(scalar, rl, cl, ret, update);
-	}
-
-	@Override
-	public MatrixBlock slice(int rl, int ru, int cl, int cu, boolean deep, CacheBlock ret) {
-		printDecompressWarning("slice");
-		MatrixBlock tmp = decompress();
-		return tmp.slice(rl, ru, cl, cu, ret);
-	}
-
-	@Override
-	public void slice(ArrayList<IndexedMatrixValue> outlist, IndexRange range, int rowCut, int colCut, int blen,
-		int boundaryRlen, int boundaryClen) {
-		printDecompressWarning("slice");
-		try {
-			MatrixBlock tmp = decompress();
-			tmp.slice(outlist, range, rowCut, colCut, blen, boundaryRlen, boundaryClen);
-		}
-		catch(DMLRuntimeException ex) {
-			throw new RuntimeException(ex);
-		}
 	}
 
 	@Override
